Day 46: 𝟮𝗗 𝗗𝗶𝗰𝘁𝗶𝗼𝗻𝗮𝗿𝗶𝗲𝘀
𝟮𝗗 𝗗𝗶𝗰𝘁𝗶𝗼𝗻𝗮𝗿𝗶𝗲𝘀
Oh yes, dictionaries are back. That can mean only one thing. It's time to make the lawyers nervous again with the return of Mokébeasts! Yay!

You'll have to wait for your beast fix though.

First, our examples will focus on the completely orignal board game of 'A murder has happened in a room with a weapon and you have to guess who did it'. 😆

Nice. Catchy.

Remember that dictionaries are very similar to lists, except that they store data as key:value pairs. The value is what it's worth and the key is what it is called. The key is used to access the value, and keys are more meaningful than index numbers.

Today we are going to expand our mad dictionary skills into the second dimension.


𝗗𝘆𝗻𝗮𝗺𝗶𝗰𝗮𝗹𝗹𝘆 𝗔𝗱𝗱𝗶𝗻𝗴 𝗧𝗼 𝗔 𝟮𝗗 𝗗𝗶𝗰𝘁𝗶𝗼𝗻𝗮𝗿𝘆
👉 This code dynamically adds to a 2D dictionary by starting with an empty dictionary and using an infinite loop to add user input.
[ ]:  clue = {}

[ ]:  while True:
[ ]:    name = input("Name: ")
[ ]:    location = input("Location: ")
[ ]:    weapon = input("Weapon: ")

[ ]:    clue[name] = {"location": location, "weapon":weapon} #line 7

[ ]:    print(clue)

The real magic happens on the 7th line of code. Instead of using .append() like we would with a list, we create a new dictionary entry.

The key is the name of the beast, but the value is a whole new dictionary that contains the details of the beast.

Each key:value pair in the dictionary is now a key that accesses a related dictionary.

Look at the output and you can see the new dictionary nested inside the first one.
👉 Can you try this code out with Lady Blurple with a rope in the study?