Day 48: 𝗦𝗮𝘃𝗶𝗻𝗴 𝘁𝗼 𝗮 𝗳𝗶𝗹𝗲
𝗦𝗮𝘃𝗶𝗻𝗴 𝘁𝗼 𝗙𝗶𝗹𝗲𝘀

👉 Now let's get some input, store it in a variable, and write it to the file.
[ ]:  f = open("savedFile.txt", "w")
[ ]:  whatText = input("> ")
[ ]:  f.write(whatText)
[ ]:  f.close()

𝙋𝙧𝙚𝙫𝙚𝙣𝙩𝙞𝙣𝙜 𝙊𝙫𝙚𝙧𝙬𝙧𝙞𝙩𝙚

We're going to change the file permissions from 'w' to 'a+'.

'a' means append - add to the end of the file.

However, if the file doen't exist, then it will crash.

'a+' means 'add to the end of the file, or create a new one if it doesn't exist'.

👉 Here's the amended code with the change on line 1:

[ ]:  f = open("savedFile.txt", "a+")
[ ]:  whatText = input("> ")
[ ]:  f.write(whatText)
[ ]:  f.close()

The problem with this is that it just glues the second input straight on to the first. Like this:
>> ## Hello thereHere's some more

So we need some.......

𝙉𝙚𝙬 𝙇𝙞𝙣𝙚𝙨

👉 We can use our old friend, the fString, to format a new line. I've used the \n new line character.

[ ]:  f = open("savedFile.txt", "a+")
[ ]:  whatText = input("> ")
[ ]:  f.write(f"{whatText}\n")
[ ]:  f.close()

Make sure you follow all three steps so the file saves.