👉 Day 49 Challenge: Highest From a File
👉 𝗗𝗮𝘆 𝟰𝟵 𝗖𝗵𝗮𝗹𝗹𝗲𝗻𝗴𝗲

In yesterday's challenge, you created a file called high.score and stored some high scores in it.

We've added a version of that file to this repl.

Your program should:

1. Read in the data from the high.score file.
2. Identify which of those users had the highest score. Automatically! (Not you doing it!)
3. Output the name and score of the winner.
Example:

🌟Current Leader🌟

Analyzing high scores......

Current leader is DJM 898,000


𝗛𝗜𝗡𝗧𝗦 💡::
a. Read each element one at a time.
b. Split each element into two pieces.
c. You'll have to cast one element as an integer to make it a number.
d. Think back to list indexing to access the second index for the score.
e. Use a max_score list to store the details of the high scorer (starting with line 1 from the file, overwrite the details if the current line has a higher score).