Day 51: 𝗔𝘂𝘁𝗼-𝗦𝗮𝘃𝗲 & 𝗔𝘂𝘁𝗼-𝗟𝗼𝗮𝗱
𝗦𝗮𝘃𝗲 𝘁𝗼 𝗮𝗻𝗱 𝗟𝗼𝗮𝗱 𝗙𝗿𝗼𝗺 𝗙𝗶𝗹𝗲𝘀

There are some things that primary storage in the RAM does better.
For example, it's easy to access, amend, or remove a piece of data held in a list (in the RAM).

Holding data in secondary storage in a file (on the disk) makes this more difficult. Or does it?

With Python, there's more than meets the eye.

👉 The program below lets me add & remove events and dates into a diary system. It adds the name & date of an event to the 2D list. Or it searches for an existing name & date and removes it.

--- [refer to the myEvents program file] ---

Manually saving and loading from this program to a file each time would be a massive faff. Instead, we can set up an auto-save by writing the save code at the end of our infinite loop.

𝗔𝘂𝘁𝗼-𝗦𝗮𝘃𝗲

👉 At the bottom of the code, we are going to add an autosave just before the loop repeats.

Make sure this new code matches the indent for the while loop, so it is part of the loop.

--- [refer to the Auto Save program file] ---

🐞 There is a potential problem with this system. Try running the program a few times and add events to the calendar.

Let's Fix That Too