Day 51: 𝗣𝗿𝗲𝘃𝗲𝗻𝘁𝗶𝗻𝗴 𝗗𝗮𝘁𝗮 𝗟𝗼𝘀𝘀
𝗣𝗿𝗲𝘃𝗲𝗻𝘁𝗶𝗻𝗴 𝗗𝗮𝘁𝗮 𝗟𝗼𝘀𝘀

Did you find the problem?

Yep. Every time we run the program, it creates a new, blank 𝐦𝐲𝐄𝐯𝐞𝐧𝐭𝐬[ ] list which gets written to the file.

This overwrites any events in the file that we saved when we ran the program previously.

To solve this, we set up the program to load any pre-existing data from the file into the myEvents list at the very start of the code.

Pay close attention to the eval() function. It's the special sauce here...
👉 𝐞𝐯𝐚𝐥( ) takes the text from the file, converts it into running code, and assigns it to 𝐦𝐲𝐄𝐯𝐞𝐧𝐭𝐬[ ] as a 2D list. Good, eh?

--- [refer to the Auto Load program file] ---


Try it out! Did you run the auto-save code first?