Day 51: 𝗖𝗼𝗺𝗺𝗼𝗻 𝗘𝗿𝗿𝗼𝗿
𝗖𝗼𝗺𝗺𝗼𝗻 𝗘𝗿𝗿𝗼𝗿

𝗡𝗼 𝗦𝘂𝗰𝗵 𝗙𝗶𝗹𝗲
Suppose you delete the file and then try to rerun the program.

[ ]:  f = open("calendar.txt","r") 
[ ]:  myEvents = eval(f.read())
[ ]:  f.close()

👉 Why am I getting a 'no such file' error?

Well friends, this happens when the file does not exist yet. The auto-load code below tries to open the file and if it can't find it, it crashes. There has to be a 'calendar.txt' file for it to work.

This fix is a temporary one because we'll learn how to sort this properly in tomorrow's lesson.

For today, we'll just comment out the auto-load code to give the auto-save the chance to create the file.

[ ]:  #f = open("calendar.txt","r") 
[ ]:  #myEvents = eval(f.read())
[ ]:  #f.close()

Once the file has been created, remove the comments to enable auto-load.