Day 52: Brace for Impact: Try and Except
𝗬𝗼𝘂 𝗮𝗿𝗲 𝗮 𝗦𝗼𝗳𝘁𝘄𝗮𝗿𝗲 𝗗𝗲𝘃𝗲𝗹𝗼𝗽𝗲𝗿!

try.... except is great for improving the user experience to reduce frustration.

However, there are problems with just putting the whole code in a 'try except'.

As developers (yes you are a software developer now), it'd be nice to know what sort of error has occurred so that we have a better idea of how to fix it.

We can tell except what type of error(s) to look for. Exception (capital 'E') means 'every type'. I've captured the error type in the 'err' variable and printed it out to tell me what the error is. Here's a list of some built in except error codes.

👉 Look at how I've extended the except now.

[ ]:  myStuff = []

[ ]:  try:
[ ]:    f.open("Stuff.mine","r")
[ ]:    myStuff = eval(f.read())
[ ]:    f.close()
[ ]:  # Try to find a file called 'Stuff.mine' and open it

[ ]:  except Exception as err:
[ ]:    print("ERROR: Unable to load")
[ ]:    print(err)
  

[ ]:  for row in myStuff:
[ ]:    print(row)